Restructure OARS → description mapping table
Make it more compact in lines of code and in memory. This introduces no
functional changes (all the mappings should be the same), but it should
make it harder to make typos when editing the table in future, and it
makes it clearer when a description is not present for an (id, value)
pair.

(This is analogous to the structure change in
872cdc8955cbf32ebb47f71f3f1108664a78913c.)